chore(native-base-components-image-slider): debugging workflows - testing
